requirements.txt
argparse
pyyaml
Babel
python-dotenv